Move testing infrastructure to be supported in downstream tests and c… (#219)
* Move testing infrastructure to be supported in downstream tests and create associated docs

* Cleanup

* Feedback updates

* Post rebase fixes

* Add fragment

* Update docs/user/design_job_testing.md


---------
