+++ changes/219.documentation
@@ -1,0 +1,1 @@
+Documented a public testing and checks framework.
