+++ nautobot_design_builder/tests/designs/context/verify_design.yaml
@@ -1,0 +1,2 @@
+---
+additional_manufacturer_3: "Manufacturer From Context YAML"
