+++ nautobot_design_builder/tests/designs/templates/verify_design.yaml.j2
@@ -1,0 +1,7 @@
+---
+manufacturers:
+    - name: "Test Manufacturer"
+    - "!create:name": "Test Manufacturer Explicit !create"
+    - "!create:name": "{{ additional_manufacturer_1 }}"
+    - "!create:name": "{{ additional_manufacturer_2 }}"
+    - "!create:name": "{{ additional_manufacturer_3 }}"
