+++ nautobot_design_builder/tests/testdata/nested_update.yaml
@@ -30,4 +30,4 @@ checks:
       - model: "nautobot.dcim.models.Interface"
         query: {name: "Ethernet1/1"}
         attribute: "description"
-      - value: "new description for Ethernet1/1"
+      - value: ["new description for Ethernet1/1"]
